backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
motor
python-dotenv
dnspython
